refactor: remove outdated license dashboard stats cron command and job
- Deleted MktLicenseDashboardStatsCronCommand and its associated files.
- Removed MktLicenseDashboardStatsCronJob and related constants.
- Introduced MktCronRegisterCommand to handle cron job registration.
- Created MktLicenseDashboardStatsRegistrationService for managing stats registration.
- Updated MktLicenseDashboardStatsService to generate and save license statistics.
- Adjusted cron pattern to run every 10 minutes.
