+++ packages/twenty-server/src/mkt-core/license/constants/mkt-license-dashboard-stats.constants.ts
@@ -1,0 +1,1 @@
+export const MKT_LICENSE_DASHBOARD_STATS_CRON_PATTERN = '*/10 * * * *'; // Every 10 minutes
