--- packages/twenty-server/src/mkt-core/license/crons/index.ts
@@ -1,12 +1,0 @@
-// Commands
-export { MktLicenseDashboardStatsCronCommand } from './commands/mkt-license-dashboard-stats.cron.command';
-
-// Jobs
-export { MktLicenseDashboardStatsCronJob } from './jobs/mkt-license-dashboard-stats.cron.job';
-
-// Constants and Types
-export {
-  MKT_LICENSE_DASHBOARD_STATS_CRON_PATTERN,
-  MktLicenseDashboardStatsCronOptions,
-  MktLicenseDashboardStatsJobData,
-} from './mkt-license-dashboard-stats.cron.constants';
