--- packages/twenty-server/src/mkt-core/license/crons/mkt-license-dashboard-stats.cron.constants.ts
@@ -1,10 +1,0 @@
-export const MKT_LICENSE_DASHBOARD_STATS_CRON_PATTERN = '0 2 * * *'; // Hàng ngày lúc 2:00 AM
-
-export interface MktLicenseDashboardStatsJobData {
-  workspaceId: string;
-}
-
-export interface MktLicenseDashboardStatsCronOptions {
-  workspaceIds?: string;
-  cronPattern?: string;
-}
